+++ RELEASE_NOTES.md
@@ -3,3 +3,4 @@ v0.93.2
 - CDSI: production enclave switch to 15637fa1
 - SVR: New enclaves for 2026Q2 for staging, and configurations (but not use) for prod.
 - node: Expose SVR2-related functionality
+- node: Support non-ASCII usernames and passwords in proxy URLs